chore(refactor): change module names for clarity
